[Ref] Clean up: Building mptrack-project should now generate much less warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@944 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/ModConvert.cpp
@@ -51,7 +51,7 @@ void UpdateEnvelopes(INSTRUMENTENVELOPE *mptEnv, CSoundFile *pSndFile, std::bits
 //---------------------------------------------------------------------------------------------------------
 {
 	// shorten instrument envelope if necessary (for mod conversion)
-	const UINT iEnvMax = pSndFile->GetModSpecifications().envelopePointsMax;
+	const uint8 iEnvMax = pSndFile->GetModSpecifications().envelopePointsMax;
 
 	#define TRIMENV(iEnvLen) if(iEnvLen > iEnvMax) { iEnvLen = iEnvMax; CHANGEMODTYPE_WARNING(wTrimmedEnvelopes); }
